Use local path in docs workflow (#89)
Rather than hardcode the URL for cloning, we can use the ${{github.repository}} variable to point to the repository. This means, if you fork the repository and setup your workflows correctly, you can get a local copy of the docs.

To generate local docs in your fork:

Click on 'actions' in the top bar and enable the workflows
Under settings -> actions -> general -> workflow permissions, set the permission to read and write
Under settings -> pages, set the branch to 'gh-pages', and click save
This should automatically generate the docs and put them in your local branch - and should update on any commits in your repo.