Default pdstools run to HC app
This PR defaults the `pdstools` CLI command to `pdstools run`. `run` is the namespace for all other apps today anyway, so if you don't supply it, there's a very high chance you actually want to go there. `pdstools help` is still there if you need it, which is a fairly standard syntax so chances are if someone is looking for help, that's what they'll try.